requirement.txt
asgiref==3.6.0   
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.6
distlib==0.3.6
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
drf-yasg==1.21.5
filelock==3.12.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
packaging==23.1
Pillow==9.5.0
pip==23.1.2
platformdirs==3.5.1
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vboxapi==1.0
vboxapi==1.0
vboxapi==1.0
virtualenv==20.23.0
